The error:
 /usr/local/pspdev/lib/gcc/psp/4.1.0/../../../../psp/include/stdint.h:84: error: conflicting types for `uint32_t`
 pspgl_misc.h:6: error: previous declaration of `uint32_t` was here

Changed 'typedef unsigned uint32_t;' to 'typedef unsigned long uint32_t;'

+++ pspgl/pspgl_misc.h
@@ -3,8 +3,7 @@
 
 #include <sys/types.h>
 
-typedef unsigned uint32_t;
-
+typedef unsigned long uint32_t;
 
 /* Return a pointer to uncached address space.  The pointer and size
    must both be a multiple CACHELINE_SIZE.  */
